modif

+++ philo_one/srcs/utils.c
@@ -6,7 +6,7 @@
 /*   By: cbussier <cbussier@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/30 15:43:27 by cbussier          #+#    #+#             */
-/*   Updated: 2020/11/19 14:10:02 by cbussier         ###   ########lyon.fr   */
+/*   Updated: 2020/11/21 14:27:01 by cbussier         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,6 +70,6 @@ void	ft_itoa(char *id, int nb)
 
 int		ft_get_timestamp(struct timeval start, struct timeval t)
 {
-	return (1000 * (t.tv_sec - start.tv_sec) +
-	(t.tv_usec - start.tv_usec) * 0.001);
+	return ((int)(float)(1000 * (t.tv_sec - start.tv_sec) +
+	(t.tv_usec - start.tv_usec) * 0.001));
 }
